ML/registration/requirement.txt
Flask
gunicorn
numpy
Pillow
dlib
Flask-HTTPAuth
"cloud-sql-python-connector[pymysql]"
google-cloud-sql-connector
sqlalchemy
google-api-python-client
google-cloud-storage